armasec-cli/pyproject.toml
[tool.poetry]
name = "armasec-cli"
version = "1.2.1"
description = "Injectable FastAPI auth via OIDC"
authors = ["Omnivector Engineering Team <info@omnivector.solutions>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/omnivector-solutions/armasec"
repository = "https://github.com/omnivector-solutions/armasec"
documentation = "https://omnivector-solutions.github.io/armasec"

packages = [
    { include = "armasec_cli"},
]

[tool.poetry.urls]
CHANGELOG = "https://github.com/omnivector-solutions/armasec/blob/main/armasec-cli/CHANGELOG.md"
CONDUCT = "https://github.com/omnivector-solutions/armasec/blob/main/CONDUCT.md"
CONTRIBUTING = "https://github.com/omnivector-solutions/armasec/blob/main/CONTRIBUTING.md"

[tool.poetry.dependencies]
python = "^3.8"
py-buzz = "^4.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
typer = "^0.9.0"
httpx = "^0.25.1"
pydantic = "^1"
snick = "^1.3.0"
auto-name-enum = "^2.0.0"
loguru = "^0.7.2"
rich = "^13.6.0"
pyperclip = "^1.8.2"

[tool.poetry.group.dev.dependencies]
ipython = ">=7,<9"
pytest = "^7.4.3"
pendulum = "^2.1.2"
plummet = {extras = ["time-machine"], version = "^1.2.1"}
pytest-random-order = "^1.1.0"
pytest-sugar = "^0.9.7"
ruff = "^0.1.5"
pytest-mock = "^3.12.0"
respx = "^0.20.2"
mypy = "^1.7.1"


[tool.poetry.scripts]
armasec = "armasec_cli.main:app"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--random-order"
testpaths = ["tests"]
asyncio_mode = "auto"

[[tool.mypy.overrides]]
module = [
    "jose",
    "jose.exceptions",
    "pyperclip",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
